requirements-dev.txt
fastapi==0.70.0
uvicorn[standard]==0.15.0
sqlalchemy==1.4.26
aiosqlite==0.17.0
greenlet==1.1.2
asyncpg==0.28.0
python-multipart==0.0.5
python-dotenv==1.0.0
pydantic==1.10.13
pytest==7.4.3
pytest-asyncio==0.23.3
httpx==0.25.2